2023.05.15 테이블 칼럼변경

+++ src/main/java/com/cook/mymealkit/mapper/CartMapper.java
@@ -11,8 +11,8 @@ public interface CartMapper {
 	
 	public void cartInsert(CartDTO dto);			// 등록
 	public List<CartDTO> cartList();				// 목록
-	public CartDTO cartFindById(int cart_id);		// 조회
+	public CartDTO cartFindById(long cart_id);		// 조회
 	public void cartUpdate(CartDTO dto);			// 수정
-	public void cartDelete(int cart_id);			// 삭제
+	public void cartDelete(long cart_id);			// 삭제
 	
 }
